Fixed CSS. Moves create user to login. cleaned up hns, js, and css files.
