--- public/css/eventstyle.css
@@ -1,169 +1,0 @@
-:root {
-  --modal_background: linear-gradient(
-    to bottom right,
-    rgb(206, 169, 105),
-    rgb(239, 223, 198)
-  );
-  --modal_font_color: rgba(230, 239, 240, 0.95);
-  --modal_input_background: rgba(230, 227, 221, 0.95);
-  --modal_input_font: rgba(52, 90, 94, 0.95);
-  --modal_input_shadow: 0 0 3px 3px rgba(128, 172, 181, 0.95);
-}
-
-.event_card {
-  color: white;
-  border: white solid;
-  border-radius: 7px;
-  padding: 5%;
-  width: 600px;
-  height: 270px;
-  margin: 5%;
-  font-family: "open sans", sans-serif;
-  font-weight: 1000;
-
-  background-size: cover;
-  transition: all 0.5s;
-}
-
-.event_card:hover {
-  width: 650px;
-  height: 290px;
-  border: solid white;
-  font-size: large;
-  box-shadow: 15px 15px 15px var(--modal_font_color); /* Add a shadow effect */
-}
-
-.detailed-event-card {
-  border: solid white;
-  margin-left: 650px;
-  /* width: 90%; */
-  height: 90%;
-
-  color: white;
-  padding: 3%;
-  background-size: cover;
-  transition: all 0.5s;
-  background-image: url(https://media.wnyc.org/i/1440/0/l/85/2020/01/DSMCandiceEvers.png);
-}
-
-.event_header {
-  display: flex;
-  border-bottom: white solid;
-  padding-bottom: 3%;
-  font-size: 150%;
-}
-
-.btnflex {
-  display: inline;
-}
-
-.event_type {
-  display: flex;
-  font-weight: bolder;
-
-  justify-content: flex-end;
-  font-size: 100%;
-}
-
-.type-detail {
-  display: flex;
-  font-weight: bolder;
-
-  justify-content: flex-end;
-}
-
-.event_date {
-  display: flex;
-  justify-content: flex-end;
-  font-size: 100%;
-}
-
-.date-detail {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.desc-detail {
-  flex: display;
-  margin: auto;
-  border: white solid;
-  width: 75%;
-}
-
-.hang-right {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.volunteer_btn {
-  border-radius: 5px;
-  color: white;
-  transition: all 0.5s;
-  height: 30px;
-  width: 50px;
-}
-
-.volunteer_btn:hover {
-  color: white;
-  font-size: larger;
-  text-decoration-line: none;
-}
-
-.footerPos {
-  align-items: baseline;
-}
-.footerBtn {
-  background-color: rgba(100, 4, 4, 0.501);
-  border-radius: 5px;
-  color: white;
-  border: white solid;
-  transition: all 0.5s;
-  height: 100px;
-  width: 150px;
-}
-
-.footerBtn:hover {
-  color: white;
-  font-size: larger;
-  text-decoration-line: none;
-}
-
-.card-style-1 {
-  background-image: url(https://t4.ftcdn.net/jpg/04/07/46/83/360_F_407468317_yxWGLvOIVyXqUrWmlPx0uePUPjkJfWeT.jpg);
-}
-
-.card-style-2 {
-  background-image: url(https://cdn.theatlantic.com/thumbor/IIbeiCxuVDtGe0CV05CSLKAyILA=/0x63:1200x738/960x540/media/img/mt/2016/05/philly/original.jpg);
-}
-
-.card-style-3 {
-  background-image: url(http://media.wbur.org/wp/2021/11/GettyImages-1298868932-1000x667.jpg);
-}
-
-#logo {
-  margin: 5%;
-  justify-content: flex-start;
-}
-
-.main-dash-btn {
-  border: solid white;
-  color: white;
-  width: 300px;
-  height: 100px;
-  background-color: rgb(76, 142, 144);
-  border-radius: 5px;
-  margin: 0 5px;
-  font-size: large;
-  align-items: center;
-}
-.main-login-btn {
-  border: solid white;
-  color: white;
-  width: 300px;
-  height: 100px;
-  background-color: rgb(76, 142, 144);
-  border-radius: 5px;
-  margin: 0 5px;
-  font-size: large;
-  align-items: center;
-}
