--- public/css/modal.css
@@ -1,78 +1,0 @@
-:root {
---modal_background: linear-gradient(to bottom right, rgb(206, 169, 105), rgb(239, 223, 198) );
---modal_font_color: rgba(230, 239, 240, 0.95);
---modal_input_background:rgba(230, 227, 221, 0.95);
---modal_input_font: rgba(52, 90, 94, 0.95);
---modal_input_shadow: 0 0 3px 3px rgba(128, 172, 181, 0.95);
-}
-
-.acctmodal , .eventmodal {
-    font-family: 'Open Sans', sans-serif;
-    display: none;
-    position: fixed; 
-    z-index: 1; 
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%,-50%);
-    width: 30%; 
-    height: 50%;
-    overflow: auto; /*for scrolling*/
-    border-radius: 5%;
-    text-align: center;
-    background-image: var(--modal_background);
-    box-shadow: 0 0 15px 4px rgba(142, 215, 230, 0.95);
-    color:var(--modal_font_color);
-    font-weight: 700;
-  
-}
-
-.modalheader {
-    font-weight: 700;
-    margin-bottom: 2%;
-    letter-spacing: 3px;
-}
-
-  .inputfields {
-    font-family: 'Open Sans' , sans-serif;
-    font-size: small;
-    background-color:var(--modal_input_background);
-    padding-left: 7px;
-    margin: 10px;
-    border: 0;
-    box-shadow: 0 0 3px 3px var(--modal_input_shadow);
-    color:var(--modal_input_font);
-    border-radius: 3px;
-    font-weight: 800;
-
-  }
-
-  .fieldplacement {
-    padding: 4%;
-  }
-
-  .submissionBtn , .eventsubmit {
-    background-color:var(--modal_input_shadow);
-    border: none;
-    padding-left: 10%;
-    padding-right: 10%;
-    margin: 2%;
-    border-radius: 3px;
-    font-weight: 700;
-    color: var(--modal_font_color) ;
-    letter-spacing: 3px;
-  }
-
-
-.label {
-    font-weight: 700;
-    letter-spacing: 3px;
-
-
-}
-
-.userclose, .eventclose {
-display: flex;
-justify-content: flex-end;
-margin: 3% 5%;
-    
-}
