+++ views/dashboard.handlebars
@@ -1,4 +1,4 @@
-    <div> <!--nav-->
+    {{!-- <div> <!--nav-->
         <nav class="bg-gradient-to-r from-blue-200 to-yellow-900">
             <!-- Brand -->
             <h1 class="ml-3 text-2xl font-semibold tracking-wide text-yellow-900 inline">
@@ -18,7 +18,7 @@
             </ul>
         
         </nav>
-        </div>
+        </div> --}}
 
 <div class="flex mb-4"><!--whole container-->
   <div class="w-1/2 bg-gradient-to-r from-blue-300 to-yellow-900 flex flex-col items-center"><!--my events contained--> 
@@ -47,48 +47,3 @@
   </div>     
 </div>
 
-<!--............MODAL.................-->
-{{!-- <div id="eventModal" class="eventmodal">
-  
-    <span class="eventclose">&times;</span>
-    
-    <h1 class="modalheader">Get a Hand</h1>
-
-<form>
-
-  <label for="event_name">event name</label>
-  <input class="inputfields" type="text" id="event_name" name="event_name"><br>
-  
-  <label for="event_date">event date</label>
-  <input class="inputfields" type="date" id="event_date" name="event_date"><br>
-  
-  <div class="dropdown">
-  <label for="event_type">event type</label><br>
-  <select name="event_type" id="event_type">
-    <option value="Yard Work">Yard Work</option>
-    <option value="Moving">Moving</option>
-    <option value="Emotional Support">Emotional Support</option>
-    <option value="Animal Care">Animal Care</option>
-    <option value="Babysitting">Babysitting</option>
-  </select>
-  </div>
-
-  <label for="volnum">number of volunteers</label>
-  <input class="inputfields" type="number" id="volnum" name="volnum"><br>
-
-    <label for="state">state</label>
-  <input class="inputfields" type="text" id="state" name="state"><br>
-
-  <label for="eventzip">zipcode</label>
-  <input class="inputfields" type="number" id="eventzip" name="eventzip"><br>
-
- <label for="event_description">event description</label>
-  <input class="inputfields" type="text" id="event_description" name="event_description"><br>
-
-<button class="submissionBtn" type="submit" id="event_submission" onclick="createEvent(event)" class="eventsubmit">create event</button> 
-  
-</form>
-
- </div>
-</div>
-<script></script> --}}
