+++ views/homepage.handlebars
@@ -1,4 +1,4 @@
-    <div> <!--nav-->
+    {{!-- <div> <!--nav-->
         <nav class="bg-gradient-to-r from-blue-200 to-yellow-900">
             <!-- Brand -->
             <h1 class="ml-3 text-2xl font-semibold tracking-wide text-yellow-900 inline">
@@ -18,7 +18,7 @@
             </ul>
         
         </nav>
-        </div>
+        </div> --}}
 
 <div class="flex flex-wrap">
