+++ views/layouts/main.handlebars
@@ -2,36 +2,55 @@
 <html lang="en">
 
 <head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>helping hands</title>
-  <link rel="stylesheet" href="./css/dashboard.css">
-  <link rel="stylesheet" href="./css/eventstyle.css">
-  <link rel="stylesheet" href="./css/modal.css">
-  {{!-- <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"> --}}
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@300&family=Poor+Story&family=Quicksand&display=swap" rel="stylesheet">
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" href="/src/styles.css">
+  <link rel="stylesheet" href="/css/styles.css">
+  <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.1/dist/jquery.slim.min.js%22%3E"></script>
+    {{!-- <script src=" https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js%22%3E"</script> --}}
+      {{!-- <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.bundle.min.js%22%3E"</script> --}}
+  <script src="https://cdn.tailwindcss.com"></script>
+
 
 </head>
 
 
-<body class="bgcolor">
-<header class="homepageNav">
-<button href="./my-event"  id= "myEventsBtn" class="homepage_btn" type="button">my events</button>
-<button type="button" id="allEventsBtn" class="all_events homepage_btn">all events</button> 
-<button  type="submit" id="logout" class="logout_btn homepage_btn">logout</button>
-</header>
+<body>
+<div> <!--nav-->
+  <nav class="bg-gradient-to-r from-blue-200 to-yellow-900">
+    <!-- Brand -->
+    <h1 class="ml-3 text-2xl font-semibold tracking-wide text-yellow-900 inline">
+      <button id="loginPage"
+        class="transition hover:scale-105 hover:no-underline hover:font-bold hover:text-yellow-900 " href="#">helping
+        hands</button>
+    </h1>
+
+    <ul class="float-right">
+      <li class="text-lg hover:text-xl tracking-wide text-blue-200 inline hover:font-semibold mr-1.5">
+        <button class="transition hover:scale-110 hover:no-underline hover:text-blue-200" id="myeventsBtn"
+          href="#"><span>my events</span></button>
+      </li>
+      <li class="text-lg hover:text-xl tracking-wide text-blue-200 inline hover:font-semibold ml-1.5 mr-1.5">
+        <button class="transition hover:scale-110 hover:no-underline hover:text-blue-200" id="allEventsBtn"
+          href="#"><span>all events</span></button>
+      </li>
+      <li class="text-lg hover:text-xl tracking-wide text-blue-200 inline hover:font-semibold ml-1.5 mr-3">
+        <button type="submit" class="transition hover:scale-110 hover:no-underline hover:text-blue-200 logout_btn homepage_btn"
+          onclick="logout()"><span>logout</span></button>
+      </li>
+    </ul>
+
+
+
+
+
+  </nav>
+</div>
 
   
-        <!-- Conditionally render login or logout links -->
-        {{#if logged_in}}
-      <button type="click" class="homepage_btn" id="myEventsBtn">welcome back!</button> 
-     {{else}}
-      {{!-- <button type="click" class="homepage_btn" id="loginPage"> go to login! </button> --}}
-        {{/if}}
   
-    <main id="mainContainer">
+    <main>
       <!-- Render the sub layout -->
       {{{ body }}}
     
@@ -46,6 +65,9 @@
   <!-- Render script for logged in users only -->
     <script src="./js/logout.js"></script>
 <script src="./js/login.js"></script>
+
+<script src="https://unpkg.com/@themesberg/flowbite@latest/dist/flowbite.bundle.js%22%3E"></script>
+
 </html>
 
 <!--.................-->
