+++ views/login.handlebars
@@ -22,10 +22,9 @@
             {{!-- <p class="text-right"><button id="signup-form" type="submit" class="text-yellow-900 text-sm font-light hover:underline"
                     href="/new-create-account">create account</button></p> --}}
         </form>
-    </div>
+    
 
-<div class="w-screen h-screen flex justify-center items-center
-    bg-gradient-to-br from-blue-200 to-amber-700">
+
   <form class="p-10 bg-white rounded-xl drop-shadow-lg space-y-5" action="">
     <h1 class="text-center text-yellow-900 text-3xl">helping hands</h1>
 
@@ -41,29 +40,13 @@
         placeholder="last name" name="lname" id="lname">
     </div>
 
-    <div class="flex flex-col space-y-2">
-      <label class="text-sm text-yellow-900" for="birthday">birthday</label>
-      <input class="w-96 px-3 py-2 rounded-md border border-yellow-900 text-yellow-900" type="date"
-        placeholder="birthday" name="birthday" id="birthday">
-    </div>
 
-    <div class="flex flex-col space-y-2">
-      <label class="text-sm text-yellow-900" for="state">state</label>
-      <input class="w-96 px-3 py-2 rounded-md border border-yellow-900 text-yellow-900" type="text" placeholder="NC"
-        name="state" id="state">
-    </div>
-
-    <div class="flex flex-col space-y-2">
-      <label class="text-sm text-yellow-900" for="userzip">zip code</label>
-      <input class="w-96 px-3 py-2 rounded-md border border-yellow-900 text-yellow-900" type="number"
-        placeholder="zip code" name="userzip" id="userzip">
-    </div>
 
     <div class="flex flex-col space-y-2">
       <label class="text-sm text-yellow-900" for="email">email</label>
-      <input class="w-96 px-3 py-2 rounded-md border border-yellow-900 text-yellow-900" type="password"
+      <input class="w-96 px-3 py-2 rounded-md border border-yellow-900 text-yellow-900" type="text"
         placeholder="email" name="email" id="email">
-    </div>
+    </div> 
 
     <div class="flex flex-col space-y-2">
       <label class="text-sm text-yellow-900" for="password">password</label>
