+++ views/partials/my-event.handlebars
@@ -6,7 +6,7 @@
             {{!-- TODO: put back format_time --}}
 
             <span id="myeventtype">your event is a {{event_type}} event on</span><span id="myeventdate">{{ event.event_date}}!</span><br>
-            <span id="myeventvol">you need {{event.vol_num}} volunteers!</span><br>
+            <span id="myeventvol">you need {{vol_num}} volunteers!</span><br>
             <span id="myeventadd">{{event_address}}</span>
 
           </p>
