--- views/specific-event-details.handlebars
@@ -1,19 +1,0 @@
-<div class="detailed-event-card">
-
-    <h1 id="speceventname">{{event.event_name}}</h1>
-    <h3 id="speceventtype">{{event.event_type}}</h3>
-{{!-- TODO: put back format_time --}}
-    <p id="speceventdate" >{{  event.event_date}}</p>
-    <p id="speceventdesc" class="desc-detail">{{event.event_description}}</p>
-    <span id="speceventstate"> {{event.event_state}}</span>, <span id="speceventzip">{{event.event_zip}} </span> 
-    <p id="specvolnum" > we need {{event.volnum}} volunteers!</p>
-    <h3 >created by: <div id="speceventusername">{{event.user.first_name}} </div> <div id="speceventuserlname"> {{event.user.last_name}} </div> </h3>
-    
-    <footer class="homepageNav flex">
-        <div class="footerPos">
-            <button id= "volBtn" class="footerBtn" type="button">give a hand</button>
-        <button id= "myEventsBtn" class="footerBtn" type="button">my
-            events</button>
-        <button id="allEventsBtn" class="footerBtn" type="button">all events</button>
-        </div>
-    </footer>
